requirements.txt
# WiFiVision - Python Dependencies
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0

# Serial communication
pyserial>=3.5

# Visualization
matplotlib>=3.4.0

# Optional: Deep learning (for advanced models)
# torch>=2.0.0  # Install separately: pip install torch --index-url https://download.pytorch.org/whl/cpu

# Optional: Computer vision (for video+CSI processing)
# opencv-python>=4.5.0
